Create getClosestLocation.html
Distances between the user and list of locations is calculated using the Haversine function

1. Define the width and height of the Google map as CSS style declaration for the attribute id'map-canvas'
2. Global declaration of the Google map objecct 'map'
3. Google DOM load event call"google.maps.event.addDomListener()' to call the Javascript function 'Init()' to obtain
the user's location, using HTML5 geolocation, to display on the map
4. Instaniate a Google LatLng coordiante object"latlng'for the latitude/longitude passed into the function"ClosestLocation()' 
from the callback function"UserLocation()'.  The location determined by the broswer is passed to "userLocation()'
in the parametner 'position'. The function'UserLocation()' then passes the latitude and longitude coordinate from the 'position' 
to 'ClosestLocation()'
5. Declare a map options parameter 'mapOptions' to set the zoom and ceter of the map
6. Instaniate(display) a map within the HTML code at the HTML tag with the attribute ind"map-canvas'
7. Declare an array'locations' to hold the coordinate data from a predefined set of locations
8. Calculate the distance between the user's location and each location in the array'locations'
For each entry in the array'location', the distance between the user's location and array entry is calculated 
using the Haversine()function. The Haversine() function is used to calculate the distance between two latitude/longitude
points on the globe. 
9. Determine which distance is the shortet between the user's location and the set of locations in the array'locations'
The variable 'mindist' will hold the current minimum(nearest)distance between the user's location and locations in the
array'locations'. It is initialized to a value greater than the maximum distance between any two points on the Earth.
The variable"closest' will hold the index into the array'locations' of the entry which is currently the minimum(nearest)distance.

For each entry in the array'location', the calculated idstance is compared to the previsous shorted distance 'mindist'. if it is 
less, then 'mindist' is set to the newest minimum distance and 'closest' is set to the index of this location' 

10. Center the map and placde a marker and InfoWindow at the nearest location

Google maps use the Google map object'google.mpas.Marker' to instantiate(display) a marker. 
which takes a marker option parameter. 

http://opengeocode.org/tutorials/googlemap/googlemaps_7.php#example_7